requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.0.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
frozenlist==1.4.0
greenlet==3.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
matplotlib-inline==0.1.7
multidict==6.0.4
nest-asyncio==1.6.0
numpy==1.26.0
obsws-python==1.6.0
openai==1.93.0
orjson==3.10.18
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytchat==0.5.5
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
SQLAlchemy==2.0.41
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.9.0
tornado==6.5.1
tqdm==4.66.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.0.4
wcwidth==0.2.13
websocket-client==1.6.3
yarl==1.9.2
zstandard==0.23.0
